+++ src/components/Expenses/Expenses.js
@@ -1,5 +1,5 @@
 import ExpenseItem from './ExpenseItem';
-import Card from './Card';
+import Card from '../UI/Card';
 import './Expenses.css';
 
 /* eslint-disable react/prop-types */
